+++ apps/client/src/features/control/playback/PlaybackControl.module.scss
@@ -2,7 +2,3 @@
   width: 100%;
   margin: 0 auto;
 }
-
-.smallButton {
-  max-width: 3em;
-}
