+++ apps/client/src/features/control/playback/extra-timer/ExtraTimer.module.scss
@@ -1,0 +1,5 @@
+.extraRow {
+  display: flex;
+  gap: 0.5rem;
+  margin-top: 2rem;
+}
